--- CodeStarterCamp_Week1/main 2.swift
@@ -1,25 +1,0 @@
-//
-//  main.swift
-//  CodeStarProject
-//
-//  Created by 유지완 on 2023/08/17.
-//
-
-import Foundation
-
-// *** 함수
-func Star(cont: Int){
-    print("***********")
-}
-
-// 막대기 함수
-func Stin(xont: Int){
-    print("    | |")
-}
-
-for cont in 1...7{
-    Star(cont: cont)
-}
-for xont in 1...4{
-    Stin(xont: xont)
-}
